[SYCL] Rely on automatic enabling of native/emulated FP atomics (#308)
After OpenCL CPU support of SPV_EXT_shader_atomic_float_* has arrived,
the compiler is switched to defining the SYCL_USE_NATIVE_FP_ATOMICS
internally for non-FPGA, non-NVPTX targets. Any further compiler changes
that may come after NVPTX/FPGA HW support lands will not require test
change, as the guarantee will remain that correct (native or emulated)
versions of FP atomic functions get emitted and executed by the HW.
